test(credentials): split the keypair vault suite into baseline and wrapping files over shared fixtures [SDK-135]
